:sparkles: Allow empty site description
Currently, site that has been previously configured with a description
can no longer be cleared out. Even if the terraform configuration
indicates empty or nil description, the change will not be persisted to
NSOT. Additionally, every time a terraform plan is run, it will always
detect an update-in-place for this field.

This change fixes this behavior by setting site description's to
space character. This is based on the pattern on other existing resource
such as network device.

Also added TestAccNetboxSite_fieldUpdate() test to test updates on
fields.

References:
- https://github.com/fbreckle/go-netbox/pull/25
- https://github.com/e-breuninger/terraform-provider-netbox/blob/master/netbox/resource_netbox_device.go#L331