Catch 404 out-of-band removal destroy
Catched a 404 not found when trying to destroy objects that were
removed out of band.
This is recommended by terraform https://developer.hashicorp.com/terraform/tutorials/providers/provider-delete

```
In addition, the destroy callback should always handle the case where the resource might already be destroyed. If the resource is already destroyed, the destroy function should not return an error.
```