fix: Add error to tags lookup if tag does not exist
There has been no error or check to determine if a tag exists already
or not. This causes the terraform to always try to make a change if the
tag does not exist, since there's no indication that the lookup didn't
find anything.

This is helpful, since it provides feedback to the user that the
associated tag does not exist, and needs to be added with the
`netbox_tag` resource.

Since there are a lot of resources that use tags, we are only activating
this feature on the netbox_virtual_machine resource to start. It can
be added to other resources by returning the diag and appending it to
the return of the function.