fix: set virtual chassis master after setting vc
When updating an existing device, it tries to set the device as a
master before updating the device to be part of a virtual chassis. This
results in the error:

```
The selected master (test-device_virtual_chassis-opff9n4aao) is not assigned to this virtual chassis.
```

Fix this by updating the device before updating the master status on the
virtual chassis.