feat: support conditions for netbox_webhook
Conditions are expected to returned as a json structure (not json
encoded string). In the provider schema, it's possible to model a map
with arbitrary typed values, so we have to provide this as a json
encoded string.

To prevent differently encoded json resulting in terraform showing
whitespace differences, we implement a DiffSuppresFunc that checks if
the old and new value are semantically equal. This mostly happens
because by default `json.Marshal()` in go omits whitespace, while
handcrafted json usually includes whitespace.